Rename: 디렉토리 및 컴포넌트 이름 변경
