+++ src/container/presetting/components/sceneSection/components/firstQuestionScene/index.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import Input from '@/components/input';
 import useStepStore from '@/container/presetting/stores/useStepStore';
 
-const QuestionScene = () => {
+const FirstQuestionScene = () => {
   const { setNextButtonOn } = useStepStore();
 
   useEffect(() => {
@@ -12,4 +12,4 @@ const QuestionScene = () => {
   return <Input>질문을 선택하세요</Input>;
 };
 
-export default QuestionScene;
+export default FirstQuestionScene;
