+++ src/container/presetting/components/sceneSection/components/interviewSettingScene/components/interviewTypeSection/index.tsx
