+++ src/container/presetting/components/sceneSection/components/interviewSettingScene/components/interviewTypeSection/type.ts
