+++ src/container/presetting/components/sceneSection/components/videoCheckScene/index.tsx
@@ -1,0 +1,19 @@
+import { MirrorView } from '@/components/camera';
+
+import useVideoCheckScene from './useVideoCheckScene';
+
+const VideoCheckScene = () => {
+  const { previewStream, isLoading, error } = useVideoCheckScene();
+
+  return (
+    <div className="mt-[1rem] h-full w-full">
+      <MirrorView
+        stream={previewStream}
+        isLoading={isLoading}
+        error={error}
+      />
+    </div>
+  );
+};
+
+export default VideoCheckScene;
